login auth handler: return correct state
If there are Liquid 2of2_no_recovery subaccounts, server may ask
to upload "required_ca" confidential addresses. Getting addresses
requires interaction with the hardware device on Liquid. However
if hardware device is empty or not on Liquid, the session is ready
to make the call, so return the correct state.