Require 2fa disable code using remaining factor
If there is more than one 2fa enabled then disabling one of them leaves
the wallet with 2fa enabled on the other factors. This is dangerous
without confirming that the user still has access to at least one of
those remaining factors, so remove the one being disabled from the list
of options for confirmation.